+++ coin-server/src/main/java/kr/or/teamserver/coinserver/controller/FileController.java
@@ -1,7 +1,6 @@
 package kr.or.teamserver.coinserver.controller;
 
 import kr.or.teamserver.coinserver.payload.UploadFileResponse;
-import kr.or.teamserver.coinserver.service.FileStorageService;
 import kr.or.teamserver.coinserver.service.FileService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
@@ -10,21 +9,26 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import javax.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/file")
 public class FileController {
 
+    private final Logger logger =  LoggerFactory.getLogger(FileController.class);
+
     @Autowired
-    private FileStorageService fileStorageService;
     private FileService fileService;
 
     @PostMapping("/upload")
     @ResponseBody
     public UploadFileResponse uploadFile(@RequestParam("file") MultipartFile file) {
-        String fileName = fileService.getFileName(file);
-        String fileDownloadUri = fileService.getFileDownloadUri(fileName);
+        String fileName = fileService.storeFile(file);
+        String fileDownloadUri = fileService.getUri(fileName);
 
         return new UploadFileResponse(fileName, fileDownloadUri,
                 file.getContentType(), file.getSize());
@@ -32,7 +36,7 @@ public class FileController {
 
     @GetMapping("/download/{fileName:.+}")
     public ResponseEntity<Resource> downloadFile(@PathVariable String fileName, HttpServletRequest request) {
-        Resource resource = fileStorageService.loadFileAsResource(fileName);
+        Resource resource = fileService.loadFileAsResource(fileName);
         String contentType = fileService.getContentType(resource, request);
 
         return ResponseEntity.ok()
