+++ coin-server/src/main/java/kr/or/teamserver/coinserver/service/FileService.java
@@ -1,29 +1,62 @@
 package kr.or.teamserver.coinserver.service;
 
-import kr.or.teamserver.coinserver.service.FileStorageService;
+import kr.or.teamserver.coinserver.exception.FileStorageException;
+import kr.or.teamserver.coinserver.exception.MyFileNotFoundException;
+import kr.or.teamserver.coinserver.property.FileStorageProperties;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
+import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import org.springframework.core.io.Resource;
-import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 
+@Service
 public class FileService {
 
-    private static final Logger logger = LoggerFactory.getLogger(FileService.class);
+    private final Path fileStorageLocation;
+    private final Logger logger = LoggerFactory.getLogger(FileService.class);
 
     @Autowired
-    private FileStorageService fileStorageService;
+    public FileService(FileStorageProperties fileStorageProperties) {
+        this.fileStorageLocation = Paths.get(fileStorageProperties.getUploadDir())
+                .toAbsolutePath().normalize();
 
-    public String getFileName(MultipartFile file) {
-        String fileName = fileStorageService.storeFile(file);
-
-        return fileName;
+        try {
+            Files.createDirectories(this.fileStorageLocation);
+        } catch (Exception ex) {
+            throw new FileStorageException("Could not create the directory where the uploaded files will be stored.", ex);
+        }
     }
 
-    public String getFileDownloadUri(String fileName) {
+    public String storeFile(MultipartFile file) {
+
+        String fileName = StringUtils.cleanPath(file.getOriginalFilename());
+
+        try {
+            if(fileName.contains("..")) {
+                throw new FileStorageException("Sorry! Filename contains invalid path sequence " + fileName);
+            }
+
+            Path targetLocation = this.fileStorageLocation.resolve(fileName);
+            Files.copy(file.getInputStream(), targetLocation, StandardCopyOption.REPLACE_EXISTING);
+
+            return fileName;
+        } catch (IOException ex) {
+            throw new FileStorageException("Could not store file " + fileName + ". Please try again!", ex);
+        }
+    }
+
+    public String getUri(String fileName) {
         String fileDownloadUri = ServletUriComponentsBuilder.fromCurrentContextPath()
                 .path("/downloadFile/")
                 .path(fileName)
@@ -32,8 +65,23 @@ public class FileService {
         return fileDownloadUri;
     }
 
+    public Resource loadFileAsResource(String fileName) {
+        try {
+            Path filePath = this.fileStorageLocation.resolve(fileName).normalize();
+            Resource resource = new UrlResource(filePath.toUri());
+            if(resource.exists()) {
+                return resource;
+            } else {
+                throw new MyFileNotFoundException("File not found " + fileName);
+            }
+        } catch (MalformedURLException ex) {
+            throw new MyFileNotFoundException("File not found " + fileName, ex);
+        }
+    }
+
     public String getContentType(Resource resource, HttpServletRequest request) {
         String contentType = "application/octet-stream";
+
         try {
             contentType = request.getServletContext().getMimeType(resource.getFile().getAbsolutePath());
         } catch (IOException ex) {
